refactor: refactor ComponentRecommanded dir.
