+++ src/main/java/org/example/techcare/dao/ComponentRecommandedDAO.java
@@ -1,6 +1,6 @@
 package org.example.techcare.dao;
 
-import org.example.techcare.model.ComponentRecommande.ComponentRecommanded;
+import org.example.techcare.model.ComponentRecommanded.ComponentRecommanded;
 import org.example.techcare.model.component.Component;
 import org.example.techcare.model.utils.ConnectionBdd;
